+++ flycs_sdk/transformations.py
@@ -1,9 +1,10 @@
 """Module containing transformations classes."""
 
+
 from enum import Enum
 from typing import List, Optional
 from flycs_sdk.custom_code import Dependency
-from flycs_sdk.query_base import QueryBase
+from flycs_sdk.query_base_schema import QueryBaseWithSchema, FieldConfig
 
 
 class WriteDisposition(Enum):
@@ -30,44 +31,6 @@ class DeltaTimeOptions(Enum):
     HOURS = "HOURS"
     DAYS = "DAYS"
     WEEKS = "WEEKS"
-
-
-class FieldConfig:
-    """FieldConfig allows to configure special options on a single field of a table."""
-
-    def __init__(self, field_name: str, decrypt: bool):
-        """Create FieldConfig object.
-
-        :param field_name: name of the field to configure
-        :type field_name: str
-        :param decrypt: whether this field should be decrypted automatically or not
-        :type decrypt: bool
-        """
-        self.field_name = field_name
-        self.decrypt = decrypt
-
-    @classmethod
-    def from_dict(cls, d: dict):
-        """Create a FieldConfig object form a dictionnary created with the to_dict method.
-
-        :param d: source dictionary
-        :type d: dict
-        :return: FieldConfig object
-        :rtype: FieldConfig
-        """
-        return FieldConfig(field_name=d["FIELD_NAME"], decrypt=d["DECRYPT"])
-
-    def to_dict(self) -> dict:
-        """Serialize the Transformation to a dictionary object.
-
-        :return: the FieldConfig as a dictionary object.
-        :rtype: Dict
-        """
-        return {"FIELD_NAME": self.field_name, "DECRYPT": self.decrypt}
-
-    def __eq__(self, other):
-        """Implement __eq__ method."""
-        return self.field_name == other.field_name and self.decrypt == other.decrypt
 
 
 class ExecutionTimeout:
@@ -108,7 +71,7 @@ class ExecutionTimeout:
         return self.delta_type == other.delta_type and self.delta == other.delta
 
 
-class Transformation(QueryBase):
+class Transformation(QueryBaseWithSchema):
     """Transformations are the lowest unit inside of a data pipeline. It is a single task implemented as a SQL query."""
 
     kind = "transformation"
@@ -118,12 +81,13 @@ class Transformation(QueryBase):
         name: str,
         query: str,
         version: str,
+        description: Optional[str] = None,
         encrypt: Optional[bool] = None,
         static: Optional[bool] = True,
         has_output: Optional[bool] = False,
         destination_table: Optional[str] = None,
         keep_old_columns: Optional[bool] = True,
-        persist_backup: Optional[bool] = True,
+        persist_backup: Optional[bool] = None,
         write_disposition: Optional[WriteDisposition] = WriteDisposition.APPEND,
         time_partitioning: Optional[dict] = None,
         cluster_fields: Optional[List[str]] = None,
@@ -138,11 +102,12 @@ class Transformation(QueryBase):
         parsing_dependencies: Optional[List[Dependency]] = None,
         destroy_table: Optional[bool] = False,
         tables: Optional[List[dict]] = None,
-        fields_config: Optional[List[FieldConfig]] = None,
         run_before_keyset: Optional[bool] = False,
         trigger_rule: Optional[str] = None,
         execution_timeout: Optional[ExecutionTimeout] = None,
         keysets_used: Optional[List[str]] = None,
+        schema: Optional[List[FieldConfig]] = None,
+        force_cache_refresh: Optional[bool] = False,
     ):
         """Class representing a transformation.
 
@@ -199,14 +164,20 @@ class Transformation(QueryBase):
         :type execution_timeout_in_minutes: in
         :param keysets_used: List of keysets used in the transformation
         :type keysets_used: List[str]
+        :param schema: List of extra configuration per field of the transformation
+        :type schema: List[FieldConfig], optional
+        :param force_cache_refresh: whether or not we need to use the cache in the pii service
+        :type force_cache_refresh: bool, optional
         """
         super().__init__(
             name=name,
             query=query,
             version=version,
+            description=description,
             encrypt=encrypt,
             static=static,
             destination_data_mart=destination_data_mart,
+            schema=schema or [],
         )
         self.has_output = has_output
         self.destination_table = destination_table
@@ -223,11 +194,11 @@ class Transformation(QueryBase):
         self.parsing_dependencies = parsing_dependencies or []
         self.destroy_table = destroy_table
         self.tables = tables
-        self.fields_config = fields_config or []
         self.run_before_keyset = run_before_keyset
         self.trigger_rule = trigger_rule
         self.execution_timeout = execution_timeout or None
         self.keysets_used = keysets_used or []
+        self.force_cache_refresh = force_cache_refresh
 
     @classmethod
     def from_dict(cls, d: dict):
@@ -242,12 +213,13 @@ class Transformation(QueryBase):
             name=d.get("NAME", ""),
             query=d["QUERY"],
             version=d["VERSION"],
+            description=d.get("DESCRIPTION", d.get("description")),
             encrypt=d.get("ENCRYPT", None),
             static=d.get("STATIC", True),
             has_output=d.get("HAS_OUTPUT", True),
             destination_table=d.get("DESTINATION_TABLE"),
             keep_old_columns=d.get("KEEP_OLD_COLUMNS", True),
-            persist_backup=d.get("PERSIST_BACKUP", True),
+            persist_backup=d.get("PERSIST_BACKUP"),
             write_disposition=WriteDisposition(
                 d.get("WRITE_DISPOSITION", "WRITE_APPEND")
             ),
@@ -266,9 +238,6 @@ class Transformation(QueryBase):
             ],
             destroy_table=d.get("DESTROY_TABLE", False),
             tables=d.get("TABLES"),
-            fields_config=[
-                FieldConfig.from_dict(x) for x in d.get("FIELDS_CONFIG") or []
-            ],
             run_before_keyset=d.get("RUN_BEFORE_KEYSET"),
             trigger_rule=d.get("TRIGGER_RULE"),
             execution_timeout=(
@@ -284,6 +253,8 @@ class Transformation(QueryBase):
                 else None
             ),
             keysets_used=d.get("KEYSETS_USED", []),
+            schema=[FieldConfig.from_dict(x) for x in d.get("SCHEMA") or []],
+            force_cache_refresh=d.get("FORCE_CACHE_REFRESH", False),
         )
 
     def to_dict(self) -> dict:
@@ -297,6 +268,7 @@ class Transformation(QueryBase):
             "NAME": self.name,
             "QUERY": self.query,
             "VERSION": self.version,
+            "DESCRIPTION": self.description,
             "ENCRYPT": self.encrypt,
             "STATIC": self.static,
             "HAS_OUTPUT": self.has_output,
@@ -318,13 +290,14 @@ class Transformation(QueryBase):
             "DESTROY_TABLE": self.destroy_table,
             "TABLES": self.tables,
             "KIND": self.kind,
-            "FIELDS_CONFIG": [config.to_dict() for config in self.fields_config],
             "RUN_BEFORE_KEYSET": self.run_before_keyset,
             "TRIGGER_RULE": self.trigger_rule,
             "EXECUTION_TIMEOUT": self.execution_timeout.to_dict()
             if self.execution_timeout
             else None,
             "KEYSETS_USED": self.keysets_used,
+            "SCHEMA": [config.to_dict() for config in self.schema],
+            "FORCE_CACHE_REFRESH": self.force_cache_refresh,
         }
 
     def __eq__(self, other):
@@ -333,6 +306,7 @@ class Transformation(QueryBase):
             self.name == other.name
             and self.query == other.query
             and self.version == other.version
+            and self.description == other.description
             and self.static == other.static
             and self.encrypt == other.encrypt
             and self.has_output == other.has_output
@@ -352,9 +326,11 @@ class Transformation(QueryBase):
             and self.destroy_table == other.destroy_table
             and self.tables == other.tables
             and self.kind == other.kind
-            and self.fields_config == other.fields_config
             and self.run_before_keyset == other.run_before_keyset
             and self.trigger_rule == other.trigger_rule
             and self.execution_timeout == other.execution_timeout
             and self.keysets_used == other.keysets_used
+            and self.schema == other.schema
+            and self.force_cache_refresh == other.force_cache_refresh
+            and self.schema == other.schema
         )