+++ tests/test_entities.py
@@ -16,6 +16,8 @@ from flycs_sdk.entities import (
 )
 from flycs_sdk.transformations import Transformation
 from flycs_sdk.views import View
+from flycs_sdk.functions import Function, Argument as FuncArg
+from flycs_sdk.procedures import StoredProcedure, Argument as ProcArg
 
 entity_name = "test"
 entity_version = "1.0.0"
@@ -38,7 +40,10 @@ class TestEntity:
             "version": entity_version,
             "kind": entity_kind,
             "stage_config": [
-                {"name": "raw", "versions": {"table_1": "1.0.0", "table_2": "1.0.0"},},
+                {
+                    "name": "raw",
+                    "versions": {"table_1": "1.0.0", "table_2": "1.0.0"},
+                },
                 {
                     "name": "staging",
                     "versions": {"table_3": "1.0.0", "table_4": "1.0.0"},
@@ -75,6 +80,40 @@ class TestEntity:
         with pytest.raises(ConflictingNameError):
             entity.add_view("staging", view)
 
+    def test_add_function(self):
+        entity = Entity(entity_name, entity_version)
+        func_arg = (FuncArg(name="argIN", type="INTEGER"),)
+        function = Function(
+            "my_function",
+            "argIN * 2",
+            "1.0.0",
+            argument_list=[func_arg],
+            return_type="INTEGER",
+        )
+        entity.add_routine("staging", function)
+        assert entity.stage_config == {"staging": {function.name: function.version}}
+        assert entity.transformations == {"staging": {function.name: function}}
+
+        with pytest.raises(ConflictingNameError):
+            entity.add_view("staging", function)
+
+    def test_add_procedure(self):
+        entity = Entity(entity_name, entity_version)
+        procedure_arg_in = (ProcArg(name="argIN", type="INTEGER", mode="IN"),)
+        procedure_arg_out = (ProcArg(name="argOUT", type="INTGER", mode="OUT"),)
+        procedure = StoredProcedure(
+            "my_procedure",
+            "BEGIN SET argOUT = argIN + argOUT; END",
+            "1.0.0",
+            argument_list=[procedure_arg_in, procedure_arg_out],
+        )
+        entity.add_routine("staging", procedure)
+        assert entity.stage_config == {"staging": {procedure.name: procedure.version}}
+        assert entity.transformations == {"staging": {procedure.name: procedure}}
+
+        with pytest.raises(ConflictingNameError):
+            entity.add_view("staging", procedure)
+
     def test_to_dict(self, my_entity):
         assert not DeepDiff(
             my_entity.to_dict(),
@@ -92,6 +131,7 @@ class TestEntity:
                         "versions": {"table_3": "1.0.0", "table_4": "1.0.0"},
                     },
                 ],
+                "location": None,
             },
             ignore_order=True,
         )
@@ -158,6 +198,7 @@ class TestBaseLayerEntity(TestEntity):
                     "versions": {"table_9": "1.0.0", "table_10": "1.0.0"},
                 },
             ],
+            "location": None,
         }
 
     def test_to_dict(self, my_entity):
@@ -189,6 +230,7 @@ class TestBaseLayerEntity(TestEntity):
                         "versions": {"table_9": "1.0.0", "table_10": "1.0.0"},
                     },
                 ],
+                "location": None,
             },
             ignore_order=True,
         )
@@ -207,6 +249,7 @@ class TestBaseLayerEntity(TestEntity):
                     {"name": "data_warehouse", "versions": {}},
                     {"name": "data_mart", "versions": {}},
                 ],
+                "location": None,
             },
             ignore_order=True,
         )
@@ -290,7 +333,10 @@ class TestNoKindEntity:
             "version": entity_version,
             "kind": None,
             "stage_config": [
-                {"name": "raw", "versions": {"table_1": "1.0.0", "table_2": "1.0.0"},},
+                {
+                    "name": "raw",
+                    "versions": {"table_1": "1.0.0", "table_2": "1.0.0"},
+                },
                 {
                     "name": "staging",
                     "versions": {"table_3": "1.0.0", "table_4": "1.0.0"},
@@ -320,6 +366,7 @@ class TestNoKindEntity:
                         "versions": {"table_3": "1.0.0", "table_4": "1.0.0"},
                     },
                 ],
+                "location": None,
             },
             ignore_order=True,
         )